debugged task 0

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -11,7 +11,7 @@ size_t print_dlistint(const dlistint_t *h)
 	if (!h)
 		return (0);
 
-	size_t nnodes;
+	int nnodes;
 	dlistint_t *curr;
 
 	while (h->prev != NULL)
